requirements.txt
pyaudio==0.2.14
numpy==1.26.3
openai-whisper==20231117
torch==2.3.1+cu118
sounddevice==0.4.6
transformers==4.42.3
